requirements.txt
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
cloudinary==1.44.1
decorator==5.2.1
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
exceptiongroup==1.3.1
executing==2.2.1
gunicorn==21.2.0
idna==3.11
ipython==8.20.0
isodate==0.7.2
jedi==0.19.2
lxml==6.0.2
matplotlib-inline==0.2.1
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.2
reportlab==4.0.9
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.3.0
zeep==4.2.1
WeasyPrint==62.3
Pillow==10.4.0
